requirements.txt
python-telegram-bot>=21.0
SQLAlchemy>=2.0
aiosqlite>=0.19
python-dotenv>=1.0
Pillow>=10.0
